upstream: update to commit 22ffb
commit 22ffb1965a0b79fbb28af5751b98814f94f6f81d
Author: Arkadiusz Bokowy <arkadiusz.bokowy@gmail.com>
Date:   Sat Nov 17 15:43:34 2018 +0100

Keep-alive time for A2DP source transport

This commit adds support for keeping A2DP transport alive for a given
number of seconds (or infinite time for negative value). One might use
this option for keeping external BT device in the powered state, or for
eliminating transport acquisition time lag when one is streaming short
audios in a repeatable matter.

Resolves #171

Change-Id: Ib388ad2bbfbe12aab1047ea44be6539671aec4de
Signed-off-by: Yu YongZhen <yuyz@rock-chips.com>